QMacStyle: fix group box for macOS Mojave
This change is based on a patch started by Gabriel. It works around the
apparently noop (on macOS Mojave, light theme)  -drawRect: call on NSBox,
replacing it with  -displayRectIgnoringOpacity:inContext:. Unfortunately,
this turns out to be  extremely slow with dark theme, so we have to resort
to a custom version of NSBox/-drawRect:  which calls its super's -drawRect:
and look more or less correct.

Task-number: QTBUG-69650
Change-Id: I5a614d7cd8002b378c6c3804663b6559e227f628
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>