doc: Add \since 6.0 to future functions
qdoc needs to know that a qdoc comment should not be part of the
documentation until a future version. In this case, some new functions
were declared to become active in Qt 6.0, but qdoc had no way of
detecting this and reported errors about them incorrectly. Adding
\since 6.0 to the qdoc comments for these functions allows qdoc to
ignore them without printing the errors.

It is also not allowed to document static functions declared in .cpp
files, because these functions are not in the public API. The qdoc
comment marker was removed from the comments for a few such static
functions.

Change-Id: I55ce0e8fb823b1dcf498d5a2436ddb20ad0a7527
Reviewed-by: Martin Smith <martin.smith@qt.io>