Fix out of bounds memory access when setting motif window hint properties
Commit b4bd5f9df3e69da707513ba544537c80a8564fb4 removed two members from
the mwm hints structure but told xcb_change_property that the structure
has still five members when it had been reduced to three.  This lead to
xcb_change_property accessing memory out of bounds.

As identified by Gatis, the safest option to avoid the access is to add
the two members again. Other window managers may be expecting their
presence in the window property.

Change-Id: Id4f0c9536cd317c35f2c6ebd1ac9ccc6f72de6a5
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>
Reviewed-by: Mikhail Svetkin <mikhail.svetkin@qt.io>