RCC: introduce compression algorithm "best"
This compression algorithm is permitted in the XML sources, which
instructs RCC to apply the best compression algorithm it has available.
If we have Zstandard available, that's its level 19 (levels 20 and up
are experimental). If not, we apply zlib compression level 9.

And apply this technique for the XDG MIME database that is built-in to
QtCore.
             Payload size  Compr. time
 Previously        313916    17.9ms
 Zlib -9           310899    53.6ms
 Zstd -14          253364    63.3ms (plus 4.0 ms on L1 heuristic check)
 Zstd -19          230647   642.5ms

Change-Id: I343f2beed55440a7ac0bfffd1562de44dbaf09cd
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>