Merge the qt_memfill{,_template} functions
We had two copies of the Duff's Device implementation, one in the .cpp
and one in the header. One of the two implementations had a protection
against zero counts, the other didn't. So move the .cpp implementation
to the header and use it in the functions that were declared there.

Fixes: QTBUG-16104
Patch-By: Allan Sandfeld Jensen <allan.jensen@qt.io>
Change-Id: Iba4b5c183776497d8ee1fffd156456cc3502946e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>