Purge some unused code for the OpenGL library name
We had a global static, a setter for it (that's nowhere used within Qt
code) and a getter for it whose only use was commented out.  Neither
was declared in any header; the getter's commented-out client had a
local extern declaration at the point of (non-)use.

Found while reviewing a change to the next few lines of code after the
commented-out use of the getter.

Change-Id: I393d56219cb7dd7cf836ca80e1bdd605a2914003
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Gatis Paeglis <gatis.paeglis@qt.io>